Point to target

+++ CameraManager/OnvifCamera/Device.cs
@@ -17,25 +17,25 @@ namespace CameraManager.OnvifCamera
         public DeviceGroup Group{ get; set; }
         public List<string> Types { get; set; }
         public List<string> Capabilities { get; set; }
-        public double Longitude { get; set; }
-        public double Latitude { get; set; }
-        public double Altitude { get; set; }
-        public double HomePanToNorth { get; set; }
-        public double HomeTiltToHorizon { get; set; }
-        public double Roll { get; set; }
-        public double Pitch { get; set; }
-        public double Yaw { get; set; }
-        public double AngleToXAxis { get; set; }
-        public double AngleToYAxis { get; set; }
-        public double AngleToZAxis { get; set; }
+        public float Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Altitude { get; set; }
+        public float HomePanToNorth { get; set; }
+        public float HomeTiltToHorizon { get; set; }
+        public float Roll { get; set; }
+        public float Pitch { get; set; }
+        public float Yaw { get; set; }
+        public float AngleToXAxis { get; set; }
+        public float AngleToYAxis { get; set; }
+        public float AngleToZAxis { get; set; }
         public bool CanPTZ { get; set; }
-        public double MinPanDegree { get; set; }
-        public double MaxPanDegree { get; set; }
-        public double MinTiltDegree { get; set; }
-        public double MaxTiltDegree { get; set; }
-        public double MinZoomLevel { get; set; }
-        public double MaxZoomLevel { get; set; }
-        public double FocalLength { get; set; }
+        public float MinPanDegree { get; set; }
+        public float MaxPanDegree { get; set; }
+        public float MinTiltDegree { get; set; }
+        public float MaxTiltDegree { get; set; }
+        public float MinZoomLevel { get; set; }
+        public float MaxZoomLevel { get; set; }
+        public float FocalLength { get; set; }
         public int Id { get; set; }
     }
 }
